requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
colorama==0.4.6
cryptography==42.0.7
dlib #Versionn 3.11 python 3.11
easygui==0.98.3
face-recognition==1.3.0
face_recognition_models==0.3.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
